add FluentResourceUpdateExample
